removed trait completely, it is not needed for now, i can finally box the hashmap and skipped serialization of ID the second type. (json prettier now)
